refactor: Extract HTTP protocol options handling into separate file
